Moved FindMojangDir from utils_minecraft.go to compatibility_windows.go and compatiblity_other_os.go to provide different implementation for different OSs
